+++ src/main/java/org/snomed/cdsservice/service/medication/MedicationConditionRuleLoaderService.java
@@ -1,4 +1,4 @@
-package org.snomed.cdsservice.service;
+package org.snomed.cdsservice.service.medication;
 
 import com.google.common.base.Strings;
 import org.apache.poi.ss.usermodel.Cell;
@@ -14,6 +14,7 @@ import org.snomed.cdsservice.model.CDSIndicator;
 import org.snomed.cdsservice.model.CDSSource;
 import org.snomed.cdsservice.model.CDSTrigger;
 import org.snomed.cdsservice.model.MedicationConditionCDSTrigger;
+import org.snomed.cdsservice.service.ServiceException;
 import org.snomed.cdsservice.service.tsclient.FHIRTerminologyServerClient;
 import org.snomed.cdsservice.util.SnomedSpreadsheetUtil;
 import org.snomed.cdsservice.util.SnomedValueSetUtil;
