+++ src/main/java/org/snomed/cdsservice/service/medication/dose/SubstanceDefinedDailyDose.java
@@ -1,0 +1,5 @@
+package org.snomed.cdsservice.service.medication.dose;
+
+public record SubstanceDefinedDailyDose(String atcRouteOfAdministration, float dose, String unit, String atcCode) {
+
+}
