+++ src/test/java/org/snomed/cdsservice/CdsServiceApplicationTests.java
@@ -1,8 +1,8 @@
 package org.snomed.cdsservice;
 
 import org.junit.jupiter.api.Test;
-import org.snomed.cdsservice.service.MedicationConditionRuleLoaderService;
-import org.snomed.cdsservice.service.MedicationCombinationRuleLoaderService;
+import org.snomed.cdsservice.service.medication.MedicationConditionRuleLoaderService;
+import org.snomed.cdsservice.service.medication.MedicationCombinationRuleLoaderService;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
